DDF-1313: Better module names
