--- spatial/geo/spatial-geo-index/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!-- /**
- * Copyright (c) Codice Foundation
- *
- * This is free software: you can redistribute it and/or modify it under the terms of the GNU Lesser General Public License as published by the Free Software Foundation, either
- * version 3 of the License, or any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- * See the GNU Lesser General Public License for more details. A copy of the GNU Lesser General Public License is distributed along with this program and can be found at
- * <http://www.gnu.org/licenses/lgpl.html>.
- *
- **/ -->
-<blueprint xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-           xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
-           xsi:schemaLocation="http://www.osgi.org/xmlns/blueprint/v1.0.0
-               http://www.osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd"
-           xmlns:cm="http://aries.apache.org/blueprint/xmlns/blueprint-cm/v1.1.0">
-
-    <bean id="geonamesIndexer" class="org.codice.ddf.spatial.geo.index.GeoNamesLuceneIndexer">
-        <cm:managed-properties persistent-id="org.codice.ddf.spatial.geo.index.properties"
-                               update-strategy="container-managed"/>
-        <property name="indexLocation" value="data/geonames-index"/>
-    </bean>
-
-    <service ref="geonamesIndexer" interface="org.codice.ddf.spatial.geo.GeoEntryIndexer"/>
-</blueprint>
