+++ spatial/spatial-app/src/main/resources/features.xml
@@ -69,18 +69,18 @@
         <bundle>mvn:org.codice.ddf.spatial/spatial-csw-endpoint/${project.version}</bundle>
     </feature>
 
-    <feature name="spatial-geo" install="manual" version="${project.version}"
+    <feature name="spatial-geocoding" install="manual" version="${project.version}"
              description="Contains the API and implementations for working with GeoNames resources">
         <bundle dependency="true">mvn:org.apache.servicemix.bundles/org.apache.servicemix.bundles.lucene/${lucene.bundle.version}</bundle>
-        <bundle>mvn:org.codice.ddf.spatial/spatial-geo-api/${project.version}</bundle>
-        <bundle>mvn:org.codice.ddf.spatial/spatial-geo-create/${project.version}</bundle>
-        <bundle>mvn:org.codice.ddf.spatial/spatial-geo-extract/${project.version}</bundle>
-        <bundle>mvn:org.codice.ddf.spatial/spatial-geo-index/${project.version}</bundle>
+        <bundle>mvn:org.codice.ddf.spatial/spatial-geocoding-api/${project.version}</bundle>
+        <bundle>mvn:org.codice.ddf.spatial/spatial-geocoding-create/${project.version}</bundle>
+        <bundle>mvn:org.codice.ddf.spatial/spatial-geocoding-extract/${project.version}</bundle>
+        <bundle>mvn:org.codice.ddf.spatial/spatial-geocoding-index/${project.version}</bundle>
     </feature>
 
     <feature name="spatial-commands" install="manual" version="${project.version}"
              description="OSGi commands">
-        <feature>spatial-geo</feature>
+        <feature>spatial-geocoding</feature>
         <bundle>mvn:org.codice.ddf.spatial/spatial-commands/${project.version}</bundle>
     </feature>
 
@@ -94,7 +94,7 @@
         <feature>spatial-wfs-v1_0_0</feature>
         <feature>spatial-wfs-v2_0_0</feature>
         <feature>spatial-csw</feature>
-        <feature>spatial-geo</feature>
+        <feature>spatial-geocoding</feature>
         <feature>spatial-commands</feature>
     </feature>
